Refactor lexicon into a split module directory
- Extracts large `LexiconEntry` static maps into `data.rs`
- Moves query logic and tests to `mod.rs`
- Resolves 'The Blob' anti-pattern while preserving public API
