+++ .jules/atlas.md
@@ -20,3 +20,12 @@
 **Tangle:** Several modules under `src/tools/` (specifically `cache`, `report`, and `ui`) and `src/semantic/assembly/` (`model`) were exposed as `pub mod`, breaking encapsulation by exposing internal implementation details to the public API.
 **Blueprint:** Modified `src/tools/mod.rs` and `src/semantic/assembly/mod.rs` to restrict these modules with `pub(crate) mod`.
 **Stability:** Achieved higher cohesion by keeping the public API surface minimal and ensuring internal structures don't leak out of their intended domains.
+
+## [Splitting The Blob: Lexicon]
+**Tangle:** `src/morphology/lexicon.rs` was a monolithic file (~2800 lines) mixing DTOs (`LexiconEntry`), a huge static `LazyLock` dictionary map of words, and helper query functions alongside massive test suites.
+**Blueprint:**
+1. Created `src/morphology/lexicon/` module directory.
+2. Extracted the massive dictionary definitions and DTOs into `src/morphology/lexicon/data.rs`.
+3. Moved the helper logic functions and tests to `src/morphology/lexicon/mod.rs`.
+4. Maintained API backward compatibility using `pub use data::*;` in `mod.rs`.
+**Stability:** Improves separation of concerns (Data vs Logic) and greatly reduces file size without breaking downstream consumers.
